refactor(client): extract NOT_LEADER classification into leader_hint module
classify_not_leader_hint and rejects_plaintext_hint now live with the cache
write they gate (compare_and_set_leader), co-locating the decision and its
side effect. attempt() and the retry loop both route NOT_LEADER through here.
Doc link in transport.rs updated.
